payments-service/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
stripe>=2.60.0
kafka-python>=2.0.2
python-dotenv>=0.19.0
requests>=2.26.0
uuid>=1.30
gunicorn==20.1.0

#pip install -r requirements.txt